+++ src/main/java/ru/skypro/diplom/repository/AdRepository.java
@@ -3,5 +3,5 @@ package ru.skypro.diplom.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.diplom.model.Ad;
 
-public interface AdRepository extends JpaRepository<Ad, Long> {
+public interface AdRepository extends JpaRepository<Ad, Integer> {
 }
