+++ src/main/java/ru/skypro/diplom/repository/AvatarRepository.java
@@ -3,6 +3,6 @@ package ru.skypro.diplom.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.diplom.model.Avatar;
 
-public interface AvatarRepository extends JpaRepository<Avatar, Long> {
+public interface AvatarRepository extends JpaRepository<Avatar, Integer> {
 
 }
